Update `emoji_str()`
The original function used to return the format of an emoji with the **first**`<` in an italicized manner and it was away from the colon by one space, which separates the symbols from the `name` of the emoji. 

This has been changed. Besides replacing "<" with "<`*`", I made replaced the "<" with an empty string.

This fixes it.